Move all cache policy-related modules to separate 'policy' subpackage
