+++ pyproject.toml
@@ -151,7 +151,6 @@ line_length = 100
 skip_gitignore = true
 skip = [
     'examples/',
-    'requests_cache/__init__.py',
     'tests/compat/',
 ]
 known_first_party = ['tests']
