+++ requests_cache/backends/base.py
@@ -17,10 +17,10 @@ from typing import Iterable, Iterator, Optional, Tuple, Union
 from requests import PreparedRequest, Response
 
 from ..cache_keys import create_key, redact_response
-from ..expiration import ExpirationTime
 from ..models import CachedResponse
+from ..policy.expiration import ExpirationTime
+from ..policy.settings import DEFAULT_CACHE_NAME, CacheSettings
 from ..serializers import init_serializer
-from ..settings import DEFAULT_CACHE_NAME, CacheSettings
 
 # Specific exceptions that may be raised during deserialization
 DESERIALIZE_ERRORS = (AttributeError, ImportError, PickleError, TypeError, ValueError)
