+++ requests_cache/models/response.py
@@ -11,11 +11,11 @@ from requests.cookies import RequestsCookieJar
 from requests.structures import CaseInsensitiveDict
 from urllib3._collections import HTTPHeaderDict
 
-from ..expiration import ExpirationTime, get_expiration_datetime
+from ..policy.expiration import ExpirationTime, get_expiration_datetime
 from . import CachedHTTPResponse, CachedRequest
 
 if TYPE_CHECKING:
-    from ..cache_control import CacheActions
+    from ..policy.actions import CacheActions
 
 DATETIME_FORMAT = '%Y-%m-%d %H:%M:%S %Z'  # Format used for __str__ only
 HeaderList = List[Tuple[str, str]]
